fixes #980_bis - apoc.load.xls Cannot convert java.util.Date to AnyValue
