+++ src/main/java/apoc/load/LoadXls.java
@@ -24,6 +24,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
+import static apoc.util.DateParseUtil.dateParse;
 import static apoc.util.Util.*;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
@@ -94,7 +95,7 @@ public class LoadXls {
         }
     }
     @Procedure("apoc.load.xls")
-    @Description("apoc.load.xls('url','selector',{config}) YIELD lineNo, list, map - load XLS fom URL as stream of row values,\n config contains any of: {skip:1,limit:5,header:false,ignore:['tmp'],arraySep:';',mapping:{years:{type:'int',arraySep:'-',array:false,name:'age',ignore:false}}")
+    @Description("apoc.load.xls('url','selector',{config}) YIELD lineNo, list, map - load XLS fom URL as stream of row values,\n config contains any of: {skip:1,limit:5,header:false,ignore:['tmp'],arraySep:';',mapping:{years:{type:'int',arraySep:'-',array:false,name:'age',ignore:false, dateFormat:'iso_date', dateParse:['dd-MM-yyyy']}}")
     public Stream<XLSResult> xls(@Name("url") String url, @Name("selector") String selector, @Name(value = "config",defaultValue = "{}") Map<String, Object> config) {
         boolean failOnError = booleanValue(config, "failOnError", true);
         try (CountingInputStream stream = FileUtils.inputStreamFor(url)) {
@@ -215,6 +216,7 @@ public class LoadXls {
 
     private static String[] convertFormat(Object value) {
         if (value == null) return null;
+        if (!(value instanceof List)) throw new RuntimeException("Only array of Strings are allowed!");
         List<String> strings = (List<String>) value;
         return strings.toArray(new String[strings.size()]);
     }
